Converter compile.js to typescript
